yet more fiddling with mysql metadata
